+++ mypy/semanal.py
@@ -8044,8 +8044,8 @@ def infer_fdef_types_from_defaults(defn: FuncDef | Decorator, self: SemanticAnal
                     typ = self.named_type("builtins.object")
                 arg_types.append(typ or UntypedType())
         ret_type = None
-        if self.options.default_return and self.options.disallow_untyped_defs:
-            ret_type = NoneType()
+        if self.options.default_return:
+            ret_type = UntypedType()
         if any(not isinstance(get_proper_type(t), AnyType) for t in arg_types) or ret_type:
             defn.type = CallableType(
                 arg_types or [UntypedType() for _ in defn.arg_kinds],
@@ -8059,19 +8059,8 @@ def infer_fdef_types_from_defaults(defn: FuncDef | Decorator, self: SemanticAnal
                 implicit=not self.options.disallow_untyped_defs,
             )
     elif defn.type:
-        assert isinstance(defn.type, CallableType)
-        if (
-            self.options.default_return
-            and is_unannotated_any(defn.type.ret_type)
-            and (
-                isinstance(defn.unanalyzed_type, CallableType)
-                and not self.options.disallow_untyped_defs
-                and any(defn.unanalyzed_type.arg_types)
-                or self.options.disallow_untyped_defs
-            )
-        ):
-            defn.type.ret_type = NoneType()
         if self.options.infer_function_types:
+            assert isinstance(defn.type, CallableType)
             for i, arg in enumerate(defn.arguments):
                 ret = None
                 if is_unannotated_any(defn.type.arg_types[i]):
@@ -8083,6 +8072,15 @@ def infer_fdef_types_from_defaults(defn: FuncDef | Decorator, self: SemanticAnal
                         defn.type.arg_types[i] = ret
                     if all(char == "_" for char in arg.variable.name):
                         defn.type.arg_types[i] = self.named_type("builtins.object")
+    if not self.options.default_return or not defn.type:
+        return
+    assert isinstance(defn.type, CallableType)
+    proper_type = get_proper_type(defn.type.ret_type)
+    if isinstance(proper_type, UntypedType):
+        if is_trivial_body(defn.body):
+            defn.type.ret_type = NoneType()
+        else:
+            proper_type.type_of_any = TypeOfAny.to_be_inferred
 
 
 def is_trivial_body(block: Block) -> bool: