+++ mypy/types.py
@@ -214,6 +214,7 @@ class TypeOfAny:
     # used to ignore Anys inserted by the suggestion engine when
     # generating constraints.
     suggestion_engine: Final = 9
+    to_be_inferred: Final = 111
 
 
 def deserialize_type(data: JsonDict | str) -> Type:
@@ -1247,6 +1248,8 @@ class AnyType(ProperType):
                 return "from a limitation"
             elif type_of_any == TypeOfAny.suggestion_engine:
                 return "from a suggestion"
+            elif type_of_any == TypeOfAny.to_be_inferred:
+                return "uninferred"
             assert False, f"unreachable: {type_of_any}"
 
         description = self.__class__.__name__.split("Type")[0]